Move commands to internal package
